requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_3cotsv2wp9/croot/certifi_1700501675943/work/certifi
dj-database-url==2.1.0
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
gunicorn==21.2.0
Markdown==3.5.2
Pillow==9.4.0
psycopg==3.1.17
psycopg-binary==3.1.17
requests==2.28.1
sqlparse @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_204sy4e5t2/croot/sqlparse_1668793448973/work
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.13
whitenoise==6.3.0

